Completed line sweep algorithm
The line sweep method is not as accurate as we thought. It recognizes
katakana and kanji with the same accuracy as it recognizes hiragana.
Furthermore, it is less accurate than previously thought, as a bug in
the code was making it check the picture to itself earlier. The number
of lines used to analyze the pictures are now drawn every 10 pixels,
yet it is still not accurate enough.